Extract interface of Property
